Fix crash when updating recipe due to incorrect ID type

+++ app/src/main/java/co/uk/bbk/culinarycompanion_francispernecita/RecipeDisplayActivity.kt
@@ -40,7 +40,7 @@ class RecipeDisplayActivity : AppCompatActivity() {
 
     override fun onResume() {
         super.onResume()
-        viewModel.observeRecipeById(recipe.id)?.observe(this) { updatedRecipe ->
+        viewModel.observeRecipeById(recipe.id.toLong())?.observe(this) { updatedRecipe ->
             if (updatedRecipe != null) {
                 recipe = updatedRecipe
                 bindRecipeDetails()
